Refactor WAL record headers into per-entry files
