+++ src/include/storage/wal/record/begin_transaction_record.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "storage/wal/record/wal_record_base.h"
+
+namespace lbug {
+namespace storage {
+
+struct BeginTransactionRecord final : WALRecord {
+    BeginTransactionRecord() : WALRecord{WALRecordType::BEGIN_TRANSACTION_RECORD} {}
+
+    void serialize(common::Serializer& serializer) const override;
+    static std::unique_ptr<BeginTransactionRecord> deserialize(common::Deserializer& deserializer);
+};
+
+} // namespace storage
+} // namespace lbug
