+++ src/include/storage/wal/record/commit_record.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "storage/wal/record/wal_record_base.h"
+
+namespace lbug {
+namespace storage {
+
+struct CommitRecord final : WALRecord {
+    CommitRecord() : WALRecord{WALRecordType::COMMIT_RECORD} {}
+
+    void serialize(common::Serializer& serializer) const override;
+    static std::unique_ptr<CommitRecord> deserialize(common::Deserializer& deserializer);
+};
+
+} // namespace storage
+} // namespace lbug
