+++ src/transaction/transaction.cpp
@@ -1,6 +1,7 @@
 #include "transaction/transaction.h"
 
 #include "catalog/catalog.h"
+#include "catalog/catalog_entry/sequence_catalog_entry.h"
 #include "common/exception/runtime.h"
 #include "main/client_context.h"
 #include "main/db_config.h"
